src/third_party/blink/web_tests/fast/html/imports/sub-imports-loop-expected.txt

CONSOLE WARNING: HTML Imports is deprecated and will be removed in M73, around March 2019. Please use ES modules instead. See https://www.chromestatus.com/features/5144752345317376 for more details.
PASS rootImport is rootImportFromChild
PASS successfullyParsed is true

TEST COMPLETE
